Updates to conics structure and functions; changing floats to doubles for precision; maybe should look into changing any such operation to be templated rather than declared explicitly

+++ include/combinatorics.h
@@ -31,13 +31,13 @@ void specialCombination(const std::vector<T> choices,
                         std::vector<std::tuple<uint, uint>>& valid_craters,
                         const N max_angle_deg) {
     const uint n = choices.size();
-    // const float max_angle_deg = 30.;
+    // const double max_angle_deg = 30.;
     const N max_angle_rad = deg2rad(max_angle_deg);
     // trig function are expensive in loops, so use the max dot product
     // angle = acos(dot(a,b)) == cos(angle) = dot(a,b)
     const N min_dot_prod = cos(max_angle_rad);
 
-    Eigen::Vector3f pt1, pt2;
+    Eigen::Vector3d pt1, pt2;
     // std::vector<std::tuple<T, T>> valid_craters;
     std::string lat1, lat2, lon1, lon2;
     T current_choice, next_choice;
@@ -63,13 +63,13 @@ template <typename T, typename N>
 std::vector<std::tuple<T, T>> specialCombination(const std::vector<T> choices, 
                                                  const N max_angle_deg) {
     const uint n = choices.size();
-    // const float max_angle_deg = 30.;
+    // const double max_angle_deg = 30.;
     const N max_angle_rad = deg2rad(max_angle_deg);
     // trig function are expensive in loops, so use the max dot product
     // angle = acos(dot(a,b)) == cos(angle) = dot(a,b)
     const N min_dot_prod = cos(max_angle_rad);
 
-    Eigen::Vector3f pt1, pt2;
+    Eigen::Vector3d pt1, pt2;
     std::vector<std::tuple<T, T>> valid_craters;
     std::string lat1, lat2, lon1, lon2;
     T current_choice, next_choice;
