+++ include/io.h
@@ -18,34 +18,34 @@ std::vector<uint> getRange(std::vector<T>);
 bool readLunarCraterEntry(  const std::string, 
                             lunar_crater&,
                             const char, 
-                            const float = 1.2, 
-                            const float = 0.9,
-                            const float = 60.0,
-                            const float = 1000.0);
+                            const double = 1.2, 
+                            const double = 0.9,
+                            const double = 60.0,
+                            const double = 1000.0);
 template <typename T>
 void runCraterReader(const std::string,
                      std::vector<T>&,
                      const char =',',
-                     const float =1.2,
-                     const float =0.9,
+                     const double =1.2,
+                     const double =0.9,
                      const uint =100);
 template <typename T>
-std::tuple<float, char> stringify_lat(const T crater);
+std::tuple<double, char> stringify_lat(const T crater);
 template <typename T>
-std::tuple<float, char> stringify_lon(const T crater);
+std::tuple<double, char> stringify_lon(const T crater);
 template <typename T>
 std::string stringify_latlon(const T crater);
-std::string stringify_latlon(const float, const float);
-std::tuple<float, char> stringify_lat(const float);
-std::tuple<float, char> stringify_lon(const float);
+std::string stringify_latlon(const double, const double);
+std::tuple<double, char> stringify_lat(const double);
+std::tuple<double, char> stringify_lon(const double);
 
 /**** Template definitions ****/
 template <typename T>
 void runCraterReader(const std::string fname,
                      std::vector<T>& craters,
                      const char sep,
-                     const float max_ell,
-                     const float min_arc,
+                     const double max_ell,
+                     const double min_arc,
                      const uint max_n_craters) {
     T db_entry;
     std::string line;
