+++ src/vector_math.cpp
@@ -1,7 +1,7 @@
 #include "vector_math.h"
 
 template <typename T>
-Eigen::Vector3f llarToWorld(const T crater, float alt, float rad) {
+Eigen::Vector3d llarToWorld(const T crater, double alt, double rad) {
 
     const T lat = crater.lat;
     const T lon = crater.lon;
@@ -9,7 +9,7 @@ Eigen::Vector3f llarToWorld(const T crater, float alt, float rad) {
     T f  = 0;                              // flattening
     T ls = atan2(pow(1 - f, 2) * tan(lat));    // lambda
 
-    Eigen::Vector3f point;
+    Eigen::Vector3d point;
     point << rad * cos(ls) * cos(lon) + alt * cos(lat) * cos(lon),
              rad * cos(ls) * sin(lon) + alt * cos(lat) * sin(lon),
              rad * sin(ls) + alt * sin(lat);
@@ -18,7 +18,7 @@ Eigen::Vector3f llarToWorld(const T crater, float alt, float rad) {
 }
 
 template <typename R, typename T>
-Eigen::Vector3f LLHtoECEF(const R crater, const T alt) {
+Eigen::Vector3d LLHtoECEF(const R crater, const T alt) {
     // see http://www.mathworks.de/help/toolbox/aeroblks/llatoecefposition.html
 
     const T lat = crater.lat;
@@ -31,7 +31,7 @@ Eigen::Vector3f LLHtoECEF(const R crater, const T alt) {
     T C      = 1/sqrt(pow(cosLat,2) + FF * pow(sinLat,2));
     T S      = C * FF;
 
-    Eigen::Vector3f point;
+    Eigen::Vector3d point;
     point << (rad * C + alt)*cosLat * cos(lon),
              (rad * C + alt)*cosLat * sin(lon),
              (rad * S + alt)*sinLat;
@@ -39,28 +39,28 @@ Eigen::Vector3f LLHtoECEF(const R crater, const T alt) {
     return point;
 }
 
-float vdot(const Eigen::Vector3f& point1, const Eigen::Vector3f& point2) {
+double vdot(const Eigen::Vector3d& point1, const Eigen::Vector3d& point2) {
     return point1.dot(point2);
 }
 
-float angularPseudodistance(const Eigen::Vector3f& point1, const Eigen::Vector3f& point2) {
+double angularPseudodistance(const Eigen::Vector3d& point1, const Eigen::Vector3d& point2) {
     return vdot(point1, point2);
 }
 
-float angularDistance(const Eigen::Vector3f& point1, const Eigen::Vector3f& point2) {
+double angularDistance(const Eigen::Vector3d& point1, const Eigen::Vector3d& point2) {
     return acos(vdot(point1, point2));
 }
 
 template <typename T>
-float latlon_dist(const T crater1, const T crater2) {
+double latlon_dist(const T crater1, const T crater2) {
     return latlon_dist(crater1.lat, crater1.lon, crater2.lat, crater2.lon);
 }
 
 template <typename T>
-float latlon_dist(const T lat1, const T lon1, const T lat2, const T lon2) {
-    Eigen::Vector3f point1 = latlon2unitVector(lat1, lon1);
+double latlon_dist(const T lat1, const T lon1, const T lat2, const T lon2) {
+    Eigen::Vector3d point1 = latlon2unitVector(lat1, lon1);
     // normalizeVector(point1);
-    Eigen::Vector3f point2 = latlon2unitVector(lat2, lon2);
+    Eigen::Vector3d point2 = latlon2unitVector(lat2, lon2);
     // normalizeVector(point2);
     return angularDistance(point1, point2);
 }
@@ -85,12 +85,12 @@ std::vector<uint> getRange(std::vector<T> vec) {
     return vec_range;
 }
 
-Eigen::Matrix3f normalizeDeterminant(const Eigen::Matrix3f& mtx) {
+Eigen::Matrix3d normalizeDeterminant(const Eigen::Matrix3d& mtx) {
     // using approach from Matrix Cookbook
     // https://www.math.uwaterloo.ca/~hwolkowi/matrixcookbook.pdf
     // get the determinant
     uint ncol = mtx.cols();
-    float det_mtx = mtx.determinant();
+    double det_mtx = mtx.determinant();
     if(abs(det_mtx)<1e-20) {
         std::cerr << "Matrix is singular/nearly singular." << std::endl;
         return mtx;
@@ -103,16 +103,16 @@ Eigen::Matrix3f normalizeDeterminant(const Eigen::Matrix3f& mtx) {
     // d = d(imag(d)==0);
     // assert that we only have one real root
     // assert(length(d)==1)
-    float d = pow(abs(1./det_mtx), 1./ncol);
+    double d = pow(abs(1./det_mtx), 1./ncol);
     // d*mtx should now have a determinant of 1
-    Eigen::Matrix3f mtx_normalized = d*mtx;
+    Eigen::Matrix3d mtx_normalized = d*mtx;
     int sign = signbit(mtx_normalized.determinant());
     // assert(abs(det(A_norm)-1)<sqrt(eps))
     return sign? -mtx_normalized : mtx_normalized;
 }
 
-Eigen::Matrix3f crossMatrix(const Eigen::Vector3f& v) {
-    Eigen::Matrix3f v_cross(3, 3);
+Eigen::Matrix3d crossMatrix(const Eigen::Vector3d& v) {
+    Eigen::Matrix3d v_cross(3, 3);
     v_cross << 0, -v(2), v(1),
               v(2), 0, -v(0),
               -v(1), v(0), 0;
